satellite-backend/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
rasterio>=1.3.0
opencv-python>=4.8.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
pillow>=10.0.0
tqdm>=4.65.0
pandas>=2.0.0
seaborn>=0.12.0
albumentations>=1.3.0
segmentation-models-pytorch>=0.3.3
google-genai>=0.2.0
python-dotenv>=1.0.0
